# shared/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateVendorDto": {
        "properties": {
          "categories": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "type": "array"
          },
          "contactEmail": {
            "format": "email",
            "type": "string"
          },
          "contactPhone": {
            "maxLength": 30,
            "type": "string"
          },
          "gstin": {
            "description": "Self-declared GSTIN; checked against the GSTIN lookup API at approval time, not at onboarding.",
            "maxLength": 15,
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "radiusKm": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "name",
          "categories"
        ],
        "type": "object"
      },
...
      "RateVendorDto": {
        "properties": {
          "comment": {
            "maxLength": 2000,
            "type": "string"
          },
          "rating": {
            "maximum": 5,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "rating"
        ],
        "type": "object"
      },
...
      "VendorAccessRequestDto": {
        "properties": {
          "purpose": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "purpose"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookie": {
        "in": "cookie",
        "name": "sid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "GateX / Society FinTech backend HTTP API, generated from the NestJS route and DTO decorators in backend/src — see backend/src/openapi/document.ts. Not hand-edited: run `npm run openapi:generate --workspace backend` to regenerate after any route/DTO change, and `npm run openapi:check --workspace backend` (wired into CI) to catch drift. Consumers generate their typed client from this file — see shared/README.md.",
    "title": "Society FinTech API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/vendors": {
      "get": {
        "operationId": "VendorsController_list",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Vendors"
        ]
      },
      "post": {
        "operationId": "VendorsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVendorDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Vendors"
        ]
      }
    },
...
    "/api/v1/vendors/residents/{residentId}/contact": {
      "get": {
        "operationId": "VendorsController_residentContact",
        "parameters": [
          {
            "in": "path",
            "name": "residentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "societyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Phase 6.3: ConsentGrant enforced at query time — see\nVendorsService.getResidentContact's doc comment. Registered before\n`GET /:id` (same reason as `GET /me` above) so \"residents\" isn't\nswallowed as an :id param. Phase 7.1: a vendor may now be linked to\nseveral societies, so the caller must name WHICH one this lookup is\nscoped to via `?societyId=` — VendorsService.getResidentContact\nvalidates the caller is actually linked to it before doing anything\nelse, so a vendor can't probe a society it has no relationship with.",
        "tags": [
          "Vendors"
        ]
      }
    },
    "/api/v1/vendors/{id}": {
      "get": {
        "operationId": "VendorsController_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Vendors"
        ]
      }
    },
    "/api/v1/vendors/{id}/access-request": {
      "post": {
        "operationId": "VendorsController_requestAccess",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorAccessRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vendors"
        ]
      }
    },
    "/api/v1/vendors/{id}/approve": {
      "post": {
        "operationId": "VendorsController_approve",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Vendors"
        ]
      }
    },
    "/api/v1/vendors/{id}/rate": {
      "post": {
        "operationId": "VendorsController_rate",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateVendorDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Vendors"
        ]
      }
    },
...
  },
  "servers": [],
...
}